+++ src/utils/mod.rs
@@ -287,13 +287,12 @@ pub(crate) mod txn_helpers {
     };
     use ark_serialize::CanonicalSerialize;
     use ark_std::{
-        collections::HashMap,
+        collections::{BTreeMap, BTreeSet},
         format,
         string::{String, ToString},
         vec,
         vec::Vec,
     };
-    use itertools::Itertools;
     use jf_primitives::merkle_tree::{MerkleLeaf, MerkleLeafProof, MerkleTree};
     use jf_utils::hash_to_field;
     use rand::{CryptoRng, RngCore};
@@ -301,7 +300,8 @@ pub(crate) mod txn_helpers {
     pub(crate) fn check_distinct_input_nullifiers(
         nullifiers: &[Nullifier],
     ) -> Result<(), TxnApiError> {
-        if nullifiers.iter().all_unique() {
+        let mut seen = BTreeSet::new();
+        if nullifiers.iter().all(|nf| seen.insert(*nf)) {
             Ok(())
         } else {
             Err(TxnApiError::InvalidParameter(
@@ -711,7 +711,7 @@ pub(crate) mod txn_helpers {
         outputs: &[&RecordOpening],
         fee: Amount,
     ) -> Result<(), TxnApiError> {
-        let mut balances = HashMap::new();
+        let mut balances = BTreeMap::new();
 
         let native_ac = inputs[0].asset_def.code; // assume already checked first is native
         balances.insert(native_ac, -(fee.0 as i128));